SetVariable to inherit directly from VT.
